user module code refactoring

+++ src/app/views/user/user-types/user-types.component.ts
@@ -12,23 +12,24 @@ import { UserTypesService } from '../../../services/user-types.service';
   styleUrls: ['./user-types.component.scss'],
 })
 export class UserTypesComponent implements OnInit {
-  modalRef: BsModalRef | null;
-  index: number;
-  message: string;
-  userTypes: any;
+  modalRef: BsModalRef | null = null;
+  index: number | null = null;
+  message: string = '';
+  userTypes: any[] = [];
   order: string = 'name';
   reverse: string = '';
-  userTypeFilter: any = '';
-  itemsForPagination: any = 5;
-  companyId: number;
+  userTypeFilter: string = '';
+  itemsForPagination: number = 5;
+  companyId: number = 0;
   globalCompany: any = {};
-  companyName: any;
-  currentRole: any;
-  highestRank: any;
-  helpFlag: any = false;
-  userName: any;
-  p: any;
+  companyName: string = '';
+  currentRole: string | null = null;
+  highestRank: any; 
+  helpFlag: boolean = false;
+  userName: string | null = null;
+  p: number = 1;
   loader = false;
+
   constructor(
     private modalService: BsModalService,
     private userTypesService: UserTypesService,
@@ -38,59 +39,62 @@ export class UserTypesComponent implements OnInit {
     private spinner: NgxSpinnerService
   ) {
     this.globalCompany = this.companyManagementService.getGlobalCompany();
-    this.companyId = this.globalCompany.companyId;
-    this.companyName = this.globalCompany.name;
+    this.companyId = this.globalCompany?.companyId ?? 0;
+    this.companyName = this.globalCompany?.name ?? '';
+
     this.getAllUserTypes();
+
     this.companyManagementService.globalCompanyChange.subscribe((value) => {
       this.globalCompany = value;
-      this.companyId = value.companyId;
-      this.companyName = value.name;
+      this.companyId = value?.companyId ?? 0;
+      this.companyName = value?.name ?? '';
       this.getAllUserTypes();
     });
   }
 
-  getAllUserTypes() {
+  ngOnInit(): void {
+    this.currentRole = sessionStorage.getItem('currentRole');
+    const highestRankStr = sessionStorage.getItem('highestRank');
+    this.highestRank = highestRankStr ? Number(highestRankStr) : null;
+
+    console.log('currentRole is', this.currentRole);
+    console.log('highestRank is', this.highestRank);
+  }
+
+  getAllUserTypes(): void {
     this.spinner.show();
-
     this.userTypes = [];
-    this.userTypesService.getAllUserTypes(this.companyId).subscribe(
-      (response) => {
-        this.spinner.hide();
 
+    this.userTypesService.getAllUserTypes(this.companyId).subscribe({
+      next: (response) => {
+        this.spinner.hide();
         this.userTypes = response;
-        const totalWarrantyTypesCount = this.userTypes.length;
-        const maxPageAvailable = Math.ceil(
-          totalWarrantyTypesCount / this.itemsForPagination
-        );
-        // Check if the current page exceeds the maximum available page
+
+        const totalUserTypesCount = this.userTypes.length;
+        const maxPageAvailable = Math.ceil(totalUserTypesCount / this.itemsForPagination);
+
         if (this.p > maxPageAvailable) {
           this.p = maxPageAvailable;
         }
-        this.userTypes.forEach((type: { parentId: any }) => {
+
+        this.userTypes.forEach((type: { parentId?: string }) => {
           if (!type.parentId) {
             type.parentId = this.companyName;
           }
         });
       },
-      (error) => {
+      error: () => {
         this.spinner.hide();
-      }
-    );
+      },
+    });
   }
 
-  ngOnInit() {
-    this.currentRole = sessionStorage.getItem('currentRole');
-    this.highestRank = sessionStorage.getItem('highestRank');
-    console.log('currentRole is' + this.currentRole);
-    console.log('highestRank is' + this.highestRank);
-  }
-
-  openModal(template: TemplateRef<any>, id: number) {
+  openModal(template: TemplateRef<any>, id: number): void {
     this.index = id;
     this.modalRef = this.modalService.show(template, { class: 'modal-lg' });
   }
 
-  closeFirstModal() {
+  closeFirstModal(): void {
     this.modalRef?.hide();
     this.modalRef = null;
   }
@@ -98,19 +102,22 @@ export class UserTypesComponent implements OnInit {
   confirm(): void {
     this.message = 'Confirmed!';
     this.spinner.show();
-
     this.userName = sessionStorage.getItem('userName');
-    this.userTypesService.removeUserType(this.index, this.userName).subscribe(
-      (response) => {
-        this.spinner.hide();
 
-        this.modalRef?.hide();
-        this.getAllUserTypes();
-      },
-      (error) => {
-        this.spinner.hide();
-      }
-    );
+    if (this.index !== null && this.userName) {
+      this.userTypesService.removeUserType(this.index, this.userName).subscribe({
+        next: () => {
+          this.spinner.hide();
+          this.modalRef?.hide();
+          this.getAllUserTypes();
+        },
+        error: () => {
+          this.spinner.hide();
+        },
+      });
+    } else {
+      this.spinner.hide();
+    }
   }
 
   decline(): void {
@@ -118,31 +125,26 @@ export class UserTypesComponent implements OnInit {
     this.modalRef?.hide();
   }
 
-  setOrder(value: string) {
+  setOrder(value: string): void {
     if (this.order === value) {
-      if (this.reverse == '') {
-        this.reverse = '-';
-      } else {
-        this.reverse = '';
-      }
+      this.reverse = this.reverse === '' ? '-' : '';
     }
     this.order = value;
   }
 
-  print() {
+  print(): void {
     this.helpFlag = false;
     window.print();
   }
 
-  help() {
+  help(): void {
     this.helpFlag = !this.helpFlag;
   }
-  onChange(e: any) {
+
+  onChange(event: any): void {
     const totalUserTypesCount = this.userTypes.length;
-    const maxPageAvailable = Math.ceil(
-      totalUserTypesCount / this.itemsForPagination
-    );
-    // Check if the current page exceeds the maximum available page
+    const maxPageAvailable = Math.ceil(totalUserTypesCount / this.itemsForPagination);
+
     if (this.p > maxPageAvailable) {
       this.p = maxPageAvailable;
     }
